added delete and edit methods
neet to fix editState and buttons

+++ src/server/trpc/router/book.ts
@@ -1,5 +1,6 @@
 import { router, publicProcedure } from "../trpc";
 import { z } from "zod";
+import { TRPCError } from "@trpc/server";
 
 export const bookRouter = router({
   getAllBooks: publicProcedure.query(({ ctx }) => {
@@ -11,6 +12,7 @@ export const bookRouter = router({
         titol: z.string(),
         autor: z.string(),
         prestatge: z.string(),
+        notes: z.string(),
       })
     )
     .mutation(({ ctx, input }) => {
@@ -18,10 +20,33 @@ export const bookRouter = router({
         data: {
           titol: input.titol,
           autor: input.autor,
-          prestatge: parseInt(input.prestatge),
+          prestatge: input.prestatge,
+          notes: input.notes,
         },
       });
     }),
+    updateBook: publicProcedure
+    .input(
+      z.object({
+        id: z.string(),
+        titol: z.string(),
+        autor: z.string(),
+        prestatge: z.string(),
+        notes: z.string(),
+      })
+    )
+    .mutation(({ ctx, input }) => {
+      return ctx.prisma.book.update({
+        where: {
+          id: input.id,
+        },
+        data: {
+          titol: input.titol,
+          autor: input.autor,
+          prestatge: input.prestatge,
+          notes: input.notes,
+        }}
+        )}),
     deleteBook: publicProcedure
     .input(
       z.object({
@@ -29,6 +54,7 @@ export const bookRouter = router({
       })
     )
     .mutation(({ ctx, input }) => {
+      if (input.id === "") throw new TRPCError({ code: "BAD_REQUEST", message: 'Cannot delete empty entry.', });
       return ctx.prisma.book.delete({
         where: {
           id: input.id,
